NavigationView - clean up and remove ItemsSourceView and replace with upstream
