+++ FluentAvalonia/UI/Controls/NavigationView/EventArgs/NavigationViewBackRequestedEventArgs.cs
@@ -2,6 +2,9 @@
 
 namespace FluentAvalonia.UI.Controls
 {
+    /// <summary>
+    /// Provides event data for the NavigationView.BackRequested event.
+    /// </summary>
     public class NavigationViewBackRequestedEventArgs : EventArgs
     {
     }
