+++ FluentAvalonia/UI/Controls/NavigationView/EventArgs/NavigationViewDisplayModeChangedEventArgs.cs
@@ -2,6 +2,9 @@
 
 namespace FluentAvalonia.UI.Controls
 {
+    /// <summary>
+    /// Provides data for the NavigationView.DisplayModeChanged event.
+    /// </summary>
     public class NavigationViewDisplayModeChangedEventArgs : EventArgs
     {
         internal NavigationViewDisplayModeChangedEventArgs(NavigationViewDisplayMode mode)
@@ -9,6 +12,9 @@ namespace FluentAvalonia.UI.Controls
             DisplayMode = mode;
         }
 
+        /// <summary>
+        /// Gets the new display mode.
+        /// </summary>
         public NavigationViewDisplayMode DisplayMode { get; }
     }
 }
