+++ FluentAvalonia/UI/Controls/NavigationView/Items/NavigationViewItemBase.cs
@@ -1,8 +1,6 @@
 ﻿using Avalonia;
 using Avalonia.Controls;
 using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace FluentAvalonia.UI.Controls
 {
