+++ FluentAvalonia/UI/Controls/NavigationView/SelectionModel/SelectionNodeOperation.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Avalonia.Controls;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
